+++ public/pages/CreateMonitor/components/DocumentLevelMonitorQueries/DocumentLevelQuery.js
@@ -36,6 +36,8 @@ const ALLOWED_DATA_TYPES = ['number', 'text', 'keyword', 'boolean'];
 // TODO DRAFT: implement validation
 export const ILLEGAL_QUERY_NAME_CHARACTERS = [' '];
 
+const DEBOUNCE_DELAY = 500;
+
 class DocumentLevelQuery extends Component {
   constructor(props) {
     super(props);
@@ -43,7 +45,10 @@ class DocumentLevelQuery extends Component {
       fieldDataType: DATA_TYPES.TEXT,
       indexFieldOptions: [],
       supportedOperators: getDocLevelQueryOperators(),
+      queryNameValue: props.query.queryName || '',
+      queryValue: props.query.query || '',
     };
+    this.debounceTimers = {};
   }
 
   componentDidMount() {
@@ -52,7 +57,32 @@ class DocumentLevelQuery extends Component {
 
   componentDidUpdate(prevProps) {
     if (prevProps.dataTypes !== this.props.dataTypes) this.initializeFieldDataType();
+    if (
+      prevProps.query.queryName !== this.props.query.queryName &&
+      this.props.query.queryName !== this.state.queryNameValue
+    ) {
+      this.setState({ queryNameValue: this.props.query.queryName });
+    }
+    if (
+      prevProps.query.query !== this.props.query.query &&
+      this.props.query.query !== this.state.queryValue
+    ) {
+      this.setState({ queryValue: this.props.query.query });
+    }
   }
+
+  componentWillUnmount() {
+    Object.values(this.debounceTimers).forEach(clearTimeout);
+  }
+
+  handleDebouncedChange = (stateKey) => (e, field, form) => {
+    const value = e.target.value;
+    this.setState({ [stateKey]: value });
+    if (this.debounceTimers[stateKey]) clearTimeout(this.debounceTimers[stateKey]);
+    this.debounceTimers[stateKey] = setTimeout(() => {
+      form.setFieldValue(field.name, value);
+    }, DEBOUNCE_DELAY);
+  };
 
   initializeFieldDataType() {
     const { dataTypes, query } = this.props;
@@ -83,6 +113,8 @@ class DocumentLevelQuery extends Component {
               inputProps={{
                 placeholder: 'Enter a name for the query',
                 isInvalid,
+                value: this.state.queryNameValue,
+                onChange: this.handleDebouncedChange('queryNameValue'),
                 'data-test-subj': `documentLevelQuery_queryName${queryIndex}`,
               }}
             />
@@ -162,6 +194,8 @@ class DocumentLevelQuery extends Component {
                   placeholder: 'Enter the search value',
                   fullWidth: true,
                   isInvalid,
+                  value: this.state.queryValue,
+                  onChange: this.handleDebouncedChange('queryValue'),
                   'data-test-subj': `documentLevelQuery_query${queryIndex}`,
                 }}
               />
@@ -180,6 +214,8 @@ class DocumentLevelQuery extends Component {
                   placeholder: 'Enter the search term',
                   fullWidth: true,
                   isInvalid,
+                  value: this.state.queryValue,
+                  onChange: this.handleDebouncedChange('queryValue'),
                   'data-test-subj': `documentLevelQuery_query${queryIndex}`,
                 }}
               />